+++ AplicationApp/Aplicacao.csproj
@@ -18,4 +18,10 @@
     <ProjectReference Include="..\PetDelivery.Exceptions\PetDelivery.Exceptions.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="UseCase\UseUsuario\Buscar\" />
+    <Folder Include="UseCase\UseUsuario\Deletar\" />
+    <Folder Include="UseCase\UseUsuario\BuscarPorEmail\" />
+  </ItemGroup>
+
 </Project>
