+++ AplicationApp/UseCase/UseProduto/Criar/IProdutoUseCase.cs
@@ -1,7 +1,7 @@
 ﻿using PetDelivery.Communication.Request;
 using PetDelivery.Communication.Response;
 
-namespace Aplicacao.UseCase.UseProduto;
+namespace Aplicacao.UseCase.UseProduto.Criar;
 public interface IProdutoUseCase
 {
 	public Task<ResponseProdutoJson> CrieProduto(RequestProdutoJson request);
