+++ AplicationApp/UseCase/UseUsuario/Criar/IRegistroDeUsuarioUseCase.cs
@@ -1,7 +1,7 @@
 ﻿using PetDelivery.Communication.Response;
 using PetDelivery.Communication.Request;
 
-namespace Aplicacao.UseCase.UserUseCase;
+namespace Aplicacao.UseCase.UseUsuario.Criar;
 public interface IRegistroDeUsuarioUseCase
 {
 	public Task<ResponseUsuarioRegistradoJson> Execute(RequestRegistroDeUsuarioJson request);
