+++ Apresentacao/Controllers/CarrinhoController.cs
@@ -19,16 +19,16 @@ public class CarrinhoController : PetDeliveryBaseController
 		return Created(string.Empty, resposta);
 	}
 
-	[HttpGet]
-	[Route("")]
-	[ProducesResponseType(typeof(ResponseCarrinhoDeComprasJson), StatusCodes.Status200OK)]
-	[ProducesResponseType(StatusCodes.Status404NotFound)]
-	public async Task<IActionResult> ObterCarrinhoAtivo(
-		[FromServices] IObterCarrinhoUseCase useCase)
-	{
-		var carrinho = await useCase.ObterCarrinhoAtivo();
+	//[HttpGet]
+	//[Route("")]
+	//[ProducesResponseType(typeof(ResponseCarrinhoDeComprasJson), StatusCodes.Status200OK)]
+	//[ProducesResponseType(StatusCodes.Status404NotFound)]
+	//public async Task<IActionResult> ObterCarrinhoAtivo(
+	//	[FromServices] IObterCarrinhoUseCase useCase)
+	//{
+	//	var carrinho = await useCase.ObterCarrinhoAtivo();
 
-		return Ok(carrinho);
-	}
+	//	return Ok(carrinho);
+	//}
 
 }
